% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partyfacts.R
\name{clear_partycolor_cache}
\alias{clear_partycolor_cache}
\title{Clear the partycoloR cache}
\usage{
clear_partycolor_cache()
}
\value{
Invisible NULL.
}
\description{
Clears cached data (Partyfacts dataset) from the current R session.
}
\examples{
clear_partycolor_cache()
}
