% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partycoloR-package.R
\docType{package}
\name{partycoloR-package}
\alias{partycoloR}
\alias{partycoloR-package}
\title{partycoloR: Extract Party Colors and Logos from Wikipedia}
\description{
Extract political party colors and logos from English Wikipedia party pages. Provides functions to scrape party infoboxes for color codes (HEX or HTML color names) and logo images. Includes integration with the Partyfacts database for easy party lookups. Designed for political scientists and party researchers working with electoral and party data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lwarode/partycoloR}
  \item Report bugs at \url{https://github.com/lwarode/partycoloR/issues}
}

}
\author{
\strong{Maintainer}: Lukas Warode \email{lukas.warode@gmx.de}

}
\keyword{internal}
