% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/party_color.R
\name{wikipedia_party_color}
\alias{wikipedia_party_color}
\title{Extract party colors from Wikipedia (legacy function)}
\usage{
wikipedia_party_color(party_url_list)
}
\arguments{
\item{party_url_list}{A list of Wikipedia political party URLs}
}
\value{
A data frame / tibble containing extracted color codes from Wikipedia
}
\description{
`r lifecycle::badge("deprecated")`

This function is deprecated. Please use [get_party_color()] instead.
}
\examples{
\dontrun{
party_list <- c(
  "https://en.wikipedia.org/wiki/Democratic_Party_(United_States)",
  "https://en.wikipedia.org/wiki/Republican_Party_(United_States)"
)
wikipedia_party_color(party_list)
}
}
