% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partycoloR-package.R
\docType{package}
\name{partycoloR-package}
\alias{partycoloR}
\alias{partycoloR-package}
\title{partycoloR: Extract Party Colors and Logos from Wikipedia}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Extract political party colors and logos from English Wikipedia party pages. Provides functions to scrape party infoboxes for color codes (HEX or HTML color names) and logo images. Includes integration with the Party Facts database for easy party lookups. Designed for political scientists and party researchers working with electoral and party data. For Party Facts, see Döring and Regel (2019) \doi{10.1177/1354068818820671} and Bederke, Döring, and Regel (2023) \doi{10.7910/DVN/TJINLQ}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lwarode/partycoloR}
  \item Report bugs at \url{https://github.com/lwarode/partycoloR/issues}
}

}
\author{
\strong{Maintainer}: Lukas Warode \email{lukas.warode@gmx.de}

}
\keyword{internal}
