% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia_party_color.R
\name{wikipedia_party_color}
\alias{wikipedia_party_color}
\title{Wikipedia party color function}
\arguments{
\item{party_url_list}{A list of Wikipedia political party URLs}
}
\value{
A data frame / tibble which contains all the extracted color codes from Wikipedia
}
\description{
Wikipedia party color function
}
\examples{
\dontrun{
party_list <- list("https://en.wikipedia.org/wiki/Democratic_Party_(United_States)",
"https://en.wikipedia.org/wiki/Republican_Party_(United_States)")
wikipedia_party_color(example_list)
}
}
